src/test/kotlin/day6/Day6SolverTest.kt
package day6

import helpers.readData
import io.kotlintest.shouldBe
import org.junit.jupiter.api.Test

internal class Day6SolverTest {
    private val daySolver = Day6Solver()
    private val dir = "day6"

    @Test
    fun `should return correct result for sample data for part1`() {
        val data = readData(dir, "sample")
        daySolver.part1(data) shouldBe 5934
    }

    @Test
    fun `should return correct result for original data for part1`() {
        val data = readData(dir, "original")
        daySolver.part1(data) shouldBe 360610
    }

    @Test
    fun `should return correct result for sample data for part2`() {
        val data = readData(dir, "sample")
        daySolver.part2(data) shouldBe 26984457539
    }

    @Test
    fun `should return correct result for original data for part2`() {
        val data = readData(dir, "original")
        daySolver.part2(data) shouldBe 1631629590423
    }
}